feat(cloning): extract cloning feature to service
